% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owdata-package.R
\docType{package}
\name{owdata-package}
\alias{owdata}
\alias{owdata-package}
\title{owdata: Index of openwashdata R Data Packages}
\description{
Index metapackage for the openwashdata R data packages. Ships a harvested catalog of every package in the openwashdata GitHub organisation, with its datasets and variables, as internal data behind accessor and search functions, and installs any listed package from the openwashdata R-universe. The catalog is refreshed by a weekly harvest.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openwashdata.github.io/owdata/}
  \item \url{https://github.com/openwashdata/owdata}
  \item Report bugs at \url{https://github.com/openwashdata/owdata/issues}
}

}
\author{
\strong{Maintainer}: Lars Schöbitz \email{lschoebitz@ethz.ch} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID})

Authors:
\itemize{
  \item Lars Schöbitz \email{lschoebitz@ethz.ch} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID})
}

Other contributors:
\itemize{
  \item Global Health Engineering, ETH Zurich [copyright holder]
}

}
\keyword{internal}
