% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owdata-package.R
\docType{package}
\name{owdata-package}
\alias{owdata}
\alias{owdata-package}
\title{owdata: Index of openwashdata R Data Packages}
\description{
An index and documentation metapackage for the openwashdata collection of R data packages. Ships harvested catalog tables of packages, datasets, and variables as package data, and provides search and install helpers together with a searchable catalog website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openwashdata/owdata}
  \item Report bugs at \url{https://github.com/openwashdata/owdata/issues}
}

}
\author{
\strong{Maintainer}: Lars Sch<U+00F6>bitz \email{lars@lse.de} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID})

}
\keyword{internal}
